moved footer and top-bar to the components folder
