+++ src/application/extension-popup/components/index.ts
@@ -1,0 +1,2 @@
+export { Footer } from './footer';
+export { TopBar } from './top-bar';
