+++ src/application/extension-popup/components/top-bar.tsx
